Share some code from AssemblyChecker.cs
