+++ src/coreclr/tools/aot/ILCompiler.Tests/TestCasesRunner/AssemblyChecker.cs
@@ -19,7 +19,7 @@ using MetadataType = Internal.TypeSystem.MetadataType;
 
 namespace Mono.Linker.Tests.TestCasesRunner
 {
-	public class AssemblyChecker
+	partial class AssemblyChecker
 	{
 		class LinkedEntity
 		{
@@ -39,7 +39,6 @@ namespace Mono.Linker.Tests.TestCasesRunner
 
 		private readonly BaseAssemblyResolver originalsResolver;
 		private readonly ReaderParameters originalReaderParameters;
-		private readonly AssemblyDefinition originalAssembly;
 		private readonly TrimmedTestCaseResult testResult;
 
 		private readonly Dictionary<AssemblyQualifiedToken, LinkedEntity> linkedMembers;
@@ -137,34 +136,6 @@ namespace Mono.Linker.Tests.TestCasesRunner
 					"Linked output includes unexpected member:\n  " +
 					string.Join ("\n  ", linkedMembers.Values.Select (e => e.Entity.GetDisplayName ())));
 		}
-
-		static bool IsCompilerGeneratedMemberName (string memberName)
-		{
-			return memberName.Length > 0 && memberName[0] == '<';
-		}
-
-		static bool IsCompilerGeneratedMember (IMemberDefinition member)
-		{
-			// Top-level methods are generated with names like
-			// <<Main>$>g__MethodName|0_1(). While the names are generated by
-			// the compiler, don't consider the method to be compiler-generated
-			// for the purpose of Kept validation, because they are attributable
-			// in source like any other method.
-			if (member is MethodDefinition method && method.Name.Contains ("<Main>$"))
-				return false;
-
-			if (IsCompilerGeneratedMemberName (member.Name))
-				return true;
-
-			if (member.DeclaringType != null)
-				return IsCompilerGeneratedMember (member.DeclaringType);
-
-			return false;
-		}
-
-		static bool IsDelegateBackingFieldsType (TypeDefinition type) => type.Name == "<>O";
-
-		static bool IsPrivateImplementationDetailsType (TypeDefinition type) => string.IsNullOrEmpty (type.Namespace) && type.Name.StartsWith ("<PrivateImplementationDetails>");
 
 		private void PopulateLinkedMembers ()
 		{
