+++ src/tools/illink/test/Mono.Linker.Tests/TestCasesRunner/AssemblyChecker.cs
@@ -15,9 +15,9 @@ using NUnit.Framework;
 
 namespace Mono.Linker.Tests.TestCasesRunner
 {
-	public class AssemblyChecker
+	partial class AssemblyChecker
 	{
-		readonly AssemblyDefinition originalAssembly, linkedAssembly;
+		readonly AssemblyDefinition linkedAssembly;
 		readonly TrimmedTestCaseResult linkedTestCase;
 
 		HashSet<string> linkedMembers;
@@ -82,30 +82,6 @@ namespace Mono.Linker.Tests.TestCasesRunner
 			}
 
 			Assert.IsEmpty (linkedMembers, "Linked output includes unexpected member");
-		}
-
-		static bool IsCompilerGeneratedMemberName (string memberName)
-		{
-			return memberName.Length > 0 && memberName[0] == '<';
-		}
-
-		static bool IsCompilerGeneratedMember (IMemberDefinition member)
-		{
-			// Top-level methods are generated with names like
-			// <<Main>$>g__MethodName|0_1(). While the names are generated by
-			// the compiler, don't consider the method to be compiler-generated
-			// for the purpose of Kept validation, because they are attributable
-			// in source like any other method.
-			if (member is MethodDefinition method && method.Name.Contains("<Main>$"))
-				return false;
-
-			if (IsCompilerGeneratedMemberName (member.Name))
-				return true;
-
-			if (member.DeclaringType != null)
-				return IsCompilerGeneratedMember (member.DeclaringType);
-
-			return false;
 		}
 
 		static bool IsBackingField (FieldDefinition field) => field.Name.StartsWith ("<") && field.Name.EndsWith (">k__BackingField");
@@ -868,12 +844,12 @@ namespace Mono.Linker.Tests.TestCasesRunner
 
 		static void VerifyPrivateImplementationDetailsType (ModuleDefinition src, ModuleDefinition linked, out TypeDefinition srcImplementationDetails, out TypeDefinition linkedImplementationDetails)
 		{
-			srcImplementationDetails = src.Types.FirstOrDefault (t => string.IsNullOrEmpty (t.Namespace) && t.Name.StartsWith ("<PrivateImplementationDetails>"));
+			srcImplementationDetails = src.Types.FirstOrDefault (t => IsPrivateImplementationDetailsType (t));
 
 			if (srcImplementationDetails == null)
 				Assert.Fail ("Could not locate <PrivateImplementationDetails> in the original assembly.  Does your test use initializers?");
 
-			linkedImplementationDetails = linked.Types.FirstOrDefault (t => string.IsNullOrEmpty (t.Namespace) && t.Name.StartsWith ("<PrivateImplementationDetails>"));
+			linkedImplementationDetails = linked.Types.FirstOrDefault (t => IsPrivateImplementationDetailsType (t));
 
 			if (linkedImplementationDetails == null)
 				Assert.Fail ("Could not locate <PrivateImplementationDetails> in the linked assembly");
@@ -1044,7 +1020,7 @@ namespace Mono.Linker.Tests.TestCasesRunner
 				return;
 
 			foreach (var nestedType in src.NestedTypes) {
-				if (nestedType.Name != "<>O")
+				if (!IsDelegateBackingFieldsType (nestedType))
 					continue;
 
 				var linkedNestedType = linked.NestedTypes.FirstOrDefault (t => t.Name == nestedType.Name);
